chore: address fb, change type, restructure

+++ internal/pkg/manifest/workload_ecs.go
@@ -87,8 +87,8 @@ type ImageWithPortAndHealthcheck struct {
 
 // AlarmArgs represents specs of CloudWatch alarms for deployment rollbacks.
 type AlarmArgs struct {
-	CPUUtilization    *int `yaml:"cpu_utilization"`
-	MemoryUtilization *int `yaml:"memory_utilization"`
+	CPUUtilization    *float64 `yaml:"cpu_utilization"`
+	MemoryUtilization *float64 `yaml:"memory_utilization"`
 }
 
 // DeploymentConfiguration represents the deployment strategies for a service.
